Auth / AuthAdmin の interface と Supabase 実装をファイル分割
- auth.ts / auth-admin.ts: interface とファクトリ関数 (getAuth / getAuthAdmin)
- supabase-auth.ts / supabase-auth-admin.ts: Supabase 用具象実装

Entire-Checkpoint: 82b437883cae
